requirements.txt
Django==2.0.7
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.5
factory-boy==2.11.1
